Refactor unit tests
Break up test.py and move the unit tests into modules that are closer to
the code that's being tested.
